feature/TASKMANAGER-SI2-16: now users are able to modify the tasks

+++ src/components/common/TaskButtonsBar/TaskButtonsBar.js
@@ -1,8 +1,6 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useCallback } from "react";
 import PropTypes from "prop-types";
 
-import Button from "@material-ui/core/Button";
-import ButtonGroup from "@material-ui/core/ButtonGroup";
 import IconButton from "@material-ui/core/IconButton";
 import DeleteIcon from "@material-ui/icons/Delete";
 import EditIcon from "@material-ui/icons/Edit";
@@ -10,31 +8,47 @@ import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 
 import { getTasks } from "../../../ducks/tasksReducer";
-import { changeModalVisibility } from "../../../ducks/modalReducer";
+import {
+  changeModalVisibility,
+  changeCreationMode,
+  changeModificationMode,
+} from "../../../ducks/modalReducer";
 import { Grid } from "@material-ui/core";
 
-const TaskButtonsBar = ({ creationMode, setModificationMode, actions, open, task }) => {
+import styles from "./styles";
+
+const TaskButtonsBar = ({ creationMode, modificationMode, actions, task }) => {
+  const classes = styles();
+
   const handleDeleteMode = useCallback(() => {
-    fetch("http://timetra.herokuapp.com/task/" + task.id, {
-      method: "DELETE",
-    })
-      .then(() => {
-        actions.getTasks();
-        actions.changeModalVisibility();
+    if (window.confirm(`¿Estas seguro que queres borrar la tarea: ${task.name}?`)) {
+      fetch("http://timetra.herokuapp.com/task/" + task.id, {
+        method: "DELETE",
       })
-      .catch((err) => console.log(err));
-  });
+        .then(() => {
+          actions.getTasks();
+          actions.changeModalVisibility();
+          actions.changeCreationMode(true);
+          actions.changeModificationMode(false);
+        })
+        .catch((err) => console.log(err));
+    }
+  }, [task, actions]);
 
   return (
     !creationMode && (
       <>
         <Grid item>
-          <IconButton color="inherit" onClick={() => setModificationMode()}>
+          <IconButton
+            className={classes.editButton}
+            onClick={() => actions.changeModificationMode(true)}
+            disabled={modificationMode}
+          >
             <EditIcon />
           </IconButton>
         </Grid>
         <Grid item>
-          <IconButton color="inherit" onClick={() => handleDeleteMode()}>
+          <IconButton className={classes.deleteButton} onClick={() => handleDeleteMode()}>
             <DeleteIcon />
           </IconButton>
         </Grid>
@@ -44,16 +58,22 @@ const TaskButtonsBar = ({ creationMode, setModificationMode, actions, open, task
 };
 
 TaskButtonsBar.propTypes = {
-  setModificationMode: PropTypes.func,
+  modificationMode: PropTypes.bool,
+  creationMode: PropTypes.bool,
+  actions: PropTypes.object,
 };
 
 const mapStateToProps = (state) => ({
-  open: state.modalReducer.open,
   task: state.modalReducer.task,
+  creationMode: state.modalReducer.creationMode,
+  modificationMode: state.modalReducer.modificationMode,
 });
 
 const mapDispatchToProps = (dispatch) => ({
-  actions: bindActionCreators({ changeModalVisibility, getTasks }, dispatch),
+  actions: bindActionCreators(
+    { changeModalVisibility, changeCreationMode, changeModificationMode, getTasks },
+    dispatch
+  ),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(TaskButtonsBar);
